+++ DesafioTapps/Property.h
@@ -2,16 +2,16 @@
 #include "Player.h"
 class Property {
 private:
-	float sellingValue;
-	float rentValue;
+	int sellingValue;
+	int rentValue;
 	int ownerId;
 	bool isSold;
 public:
-	Property(float sellingValue, float rentValue);
+	Property(int sellingValue, int rentValue);
 	Property();
 
-	float GetRentValue();
-	float GetSellingValue();
+	int GetRentValue();
+	int GetSellingValue();
 	int GetOwnerId();
 	void SetForSelling();
